Allow to configure an info URL that is shown if an OWNERS file is invalid
If an operation fails because an OWNERS file is invalid / non-parsable
we return '409 Conflict' with an error message telling which OWNERS file
is invalid and which issues it has. Since OWNERS files are maintained by
the project team, it's the responsibility of the project team, project
owners or host admins to fix the OWNERS file, but it depends on the
project who exactly is responsible for this. The new info URL that can
be configured will be included into the error message so that users can
find project-specific information about how they can report / fix
invalid OWNERS files.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I4f14842a283e478cf503b8f18d18a3643c994373