GetOwnedPath: Limit the number of returned paths
Computing all owned paths can be expensive for large changes.

Often callers do not need all owned paths, hence it should be possible
to limit the result. E.g. the frontend may use this REST endpoint to
highlight files on the change screen that are owned by the current user,
but for large changes the frontend is not showing all files by default,
hence retrieving all owned files in this case would be wasteful.

To avoid unintended high latency when using this REST endpoint we limit
the results in the response by default to 50 and if callers want to get
more results they need to set a higher limit explicitly. This behaviour
is consistent with the get code owners REST endpoints with also have a
default limit.

We also add a start parameter that allows to skip owned files. This can
be used to paginate over the owned files.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I6340d16e9a95c57a338f2f025793642305244ed1