GetCodeOwnerStatus: Limit the number of returned statuses by default
Computing all code owner statuses can be expensive for large changes.

To avoid unintended high latency when using this REST endpoint we limit
the results in the response by default to 200 and if callers want to get
more results they need to set a higher limit explicitly. This behaviour
is consistent with the get code owners REST endpoints which also have a
default limit.

By default the change screen is showing at most 200 files and if the
user wants to see more/all files they need to expand the file list. We
choose 200 as default limit for the code owner statuses so that these
200 files can be covered (renames are reported as a single code owner
status, hence also 200 renames are covered).

Having this default limit means that the frontend needs to be adapted to
request further code owner statuses when the user expands the file list
on the change screen to all files. In this case the frontend can use the
start parameter to skip the 200 code owner statuses that it already has.

While the frontend is not adapted yet, code owner statuses are not
available from the 201st file onwards when the user expands the file
list.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I73508bc23c699a4f91e70f9da041806ba7a19ce1