GetCodeOwnerStatus: Limit the number of returned statuses
Computing all code owner statuses can be expensive for large changes.

Often callers do not need all code owner statuses, hence it should be
possible to limit the result. E.g. the frontend uses this REST endpoint
to show the code owner statuses in the file list on the change screen,
but for large changes the frontend is not showing all files by default,
hence retrieving all code owner statuses in this case would be wasteful.

To avoid unintended high latency when using this REST endpoint we limit
the results in the response by default to 200 and if callers want to get
more results they need to set a higher limit explicitly. This behaviour
is consistent with the get code owners REST endpoints which also have a
default limit.

By default the change screen is showing at most 200 files and if the
users wants to see more/all files they need to expand the file list. We
choose 200 as default limit for the code owner statuses so that these
200 files can be covered (renames are reported as a single code owner
status, hence also 200 renames are covered).

Having this default limit means that the frontend needs to be adapted to
request further code owner statuses when the user expands the file list
on the change screen to all files. In this case the frontend can use the
start parameter to skip the 200 code owner statuses that it already has
by using the start parameter.

The start parameter allows to skip file code owner statuses and can be
used to paginate over the code owner statuses.

While the frontend is not adapted yet, code owner statuses are not
available from the 201st file onwards when the user expands the file
list.

Change-Id: I80acb3d1ea04d13db72700623c72dd8c4bf5f63a
Signed-off-by: Edwin Kempin <ekempin@google.com>