ChangeCodeOwners: Add request object for computing the code owner status
This allows us in follow-up changes to add parameters to the code owner
status request (e.g. a limit option to limit the number of file code
owner statuses that should be returned).

This is a backwards incompatible change for the extension API, but since
there is no other implementor of this API, it's fine to change it.

Change-Id: Ie73563dbdf887cef5fefbc943475ab31bbc7d9f2
Signed-off-by: Edwin Kempin <ekempin@google.com>