CodeOwnerApprovalCheckTest: Fix test for non-pure revert
The test intends to amend the created revert commit, but since it
forgot to fetch it, it amended the commit that is being reverted and
uploaded it as a new patch set to the revert change. This means that
this commit has a wrong parent commit. For the test it's not really
making any difference at the moment, but we rather fix this.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I0ca694ee587ed93cecda9f1c44e42e460d7d2a61