CodeOwnersPluginConfigSnapshotTest: Fix flaky tests (second attempt)
The tests were flaky because the order of elements in sets is not fixed.

The first attempt to fix this was change Id758c5e4d which sorted the
override approvals in the tests before asserting them. However this
change missed to adapt one test.

Since doing the sorting in each test is error-prone and repetitive we
now do the sorting in
CodeOwnersPluginConfigSnapshot#getOverrideApprovals() and return an
ImmutableSortedSet. This way every caller can rely on the sorting.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I38415ec45ea309b42fc5e7ed1cd98ca4c330fdd6