Set a default limit for the in-memory code owner config cache
For Android and Chrome at Google p95 is mostly below 50 code owner
config reads for change. However the worst peek that we observed for p99
is ~6.3k code owner config reads per change. Set a default limit of 25k
for the cache that is well above the worst peek. The only purpose of
this limit is to ensure that we are not running into out-of-memory
errors if a projects has a crazy amount of OWNERS files. If the limit is
reached the latency for code owner computations goes down, but that's
better then running out of memory.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Iad98f176b1aea558294e89ee7995a79827bb5bfc