Fix sorting in code owner suggestion if file is owned by all users
If an OWNERS file assigns code ownership to all users and a concrete
user, we didn't include any distance score for the concrete user. This
was because we returned early from the callback when a file is owned by
all users and the step to add distance scores for the concrete user was
skipped.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I80e99fa175411a2ce11bf409600115fc69c142ee