Add distance scores for code owners that are added for resolving all users
If --resolve-all-users is specified when listing code owners code
ownerships that are assigned to all users (aka '*') are resolved to
random users until the limit is reached. For these users no distance
score was added.

After adding random users to fill up the limit, no further code owners
will be included in the result (because the limit is reached). This
means the random users will always have the lowest score. For the
sorting this means that this fix makes no difference (since lowest score
and no score will result in the same sort position). Due to this we
can't test this fix via the sort order. Since we also don't return the
scores to the client, it's hard to test this and hence this change
doesn't include a test.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: If579300d1af822c997344ca350a93a220047f2c8