AbstractGetCodeOwnersForPath: Cleanup how subclasses add further scorings
To add further scorings subclasses needed to override the sortCodeOwners
methid to replace the CodeOwnerScorings instance. How this worked wasn't
obvious. Improve the readability by adding a getCodeOwnerScorings(...)
method that allows subclasses to contribute CodeOwnerScoring's at the
moment when the CodeOwnerScorings instance is created. This way the
sortCodeOwners method no longer needs to be overridden.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I2486c293f3f7f0dbccdc6d2231d13020a413f5de