CodeOwnerApprovalCheck: Avoid passing implicit approval flag into all methods
This was a bit confusing:

* if enableImplicitApprovalFromUploader == true, then patchSetUploader
  was guaranteed to be non-null
* if enableImplicitApprovalFromUploader == false, then patchSetUploader
  could have been null (e.g. for calls done from
  getFileStatusesForAccount(...))

Instead of passing in enableImplicitApprovalFromUploader and the
patchSetUploader, just pass in a nullable implicitApprover (account ID
of the implicit approver). If implicitApprover is set, there is an
implicit approval from that account. If implicitApprover is not set
there is no implicit approval. That's easier to understand.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ia1f102d5e9621833cb8b9723fed658cec83698fd