Load accounts in parallel when resolving code owners
After reading code owner emails from an OWNERS file they are resolved to
accounts. Resolving an email to an account requires doing multiple
checks and lookups. Doing the lookups for all emails at once is more
efficient than iterating over the emails and doing the lookups once for
each email:

1. Lookup external IDs for emails:
   The external IDs are retrieved from the external ID cache. Each
   access to the external ID cache requires reading the SHA1 of the
   refs/meta/external-ids branch. If external IDs for multiple emails
   are needed it is more efficient to use the
   ExternalIds.byEmails(String...) method instead of the
   ExternalIds.byEmail(String) method as the
   ExternalIds.byEmails(String…) method reads the SHA1 of the
   refs/meta/external-ids branch only once (and not once per email).

2. Lookup accounts:
   Accounts are retrieved from the account cache. Using the
   AccountCache.get(Set<Account.Id>) method loads accounts in parallel
   and hence it is more efficient than calling
   AccountCache.get(Account.Id) once for each account.

Doing the lookups only once for all emails requires rewriting the
CodeOwnerResolver class, as now all checks and transformations need to
be performed on the collection of emails, rather than on a single email.

To be able to transform a Map.Entry in a stream of Map.Entry's
efficiently we need to have an instantiatable Map.Entry implementation.
Since none is provided by Java, we added an own class for this (see
class ImmutableMapEntry).

While we are at it slightly improve 2 of the debug messages.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ie7dd6f54a0c24c2c4a9018d32ddb66b8f466d32a