Document another case where no code owners are suggested
Code owners that cannot see the change, and hence cannot approve it, are
filtered out from the code owner suggestion. This means if none of the
code owners can see the change, the list of suggested code owners is
empty. In this case the frontend is showing an info icons that links to
the documentation that explains the cases when this happens. Update this
documentation to include the case where none of the code owners can see
the change.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I705fc962a43f423fc0950e80096effae5a270608