Fix: this.codeOwnersCacheApi.getStatus is not a function
The code-owners plugin requests owners status for the latest patchset.
Usually, the patchset in the response matches the latest patchset
known to the Gerrit itself. However, there are 2 rare cases:

1. A new patchset is uploaded right after the gerrit loads change info,
but before the plugin requests owners status

2. A new patchset has been uploaded, but owners plugin receives
status for a previous patchset.

Before the fix, Gerrit was tried to request owner status again and
failed due to the bug in the code.

With this fix, the behavior is changed:
- In the 1st case, the plugin notifies a user about a newer patchset
and doesn't show status information until user reloads a change
- In the 2nd case, the plugin retries status request again and again
until the status for the newer patchset is recevied.

Change-Id: I9971ce76f6ffcfccfcfa69d791c92a0e24a0f1c8