Cache resolved code owners for the time of the request
When computing code owner file statuses for a change (e.g. to compute
the results for the code owners submit rule) emails that are mentioned
in relevant code owner config files need to be resolved to Gerrit
accounts. Now the resolved code owners are cached in memory for the time
of the request so that this resolution, which involves many steps, has
to be done only once per email (the same email may be mentioned multiple
times in different relevant OWNERS files).

The caching is implemented in CodeOwnerResolver. This means to have an
effect callers like CodeOwnerApprovalCheck need to use the same
CodeOwnerResolver instance to resolve code owners for all relevant
OWNERS files.

To get an understanding of much the new cache helps we add new metrics
to record how many code owners are resolved in total and per change and
how many cache reads there are (also in total and per change).

By default the cache is limited to hold at most 1000 entries per request
(1000 entries consumes ~0.124 MB). It's not a problem if the cache
limited is exceeded as in this case code owners will just be resolved
newly as before, which will result in a slightly higher latency. Still
we want to know when this happens, which is why we log a warning in this
case.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ie5e1d69fc763516f46dfee990f82420f5168a0f0