Test that change can be submitted with override if OWNERS file is non-parsable
If a relevant OWNERS file is non-parsable it's not possible to submit
the change since the code owners are unknown. In this case it should be
possible to make the change submittable by applying a code owner
override. If an override is present the code owners submission check is
satisfied and we do not need to know the code owners. This works but we
didn't have a test for this yet.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I5ff537cbacc62ed35ba3f383b468466c9398ae1c