Add a debug option for the list code owners REST endpoints
Users frequently ask why a user is or isn't suggested as a code owner.
The new debug option on the list code owners REST endpoint includes
debug logs into the response that help to answer exactly this question.

We already have the CheckCodeOwner REST endpoint which exposes similar
debug logs, but only for a single user. With the new option this
information can be retrieved for all relevant users at once.

Since debug logs may contain sensitive information this option is only
available for users that have the 'Check Code Owner' global capability,
which is also required by the CheckCodeOwner REST endpoint.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I183907365dab02732ff67732a9d09ae95eb10459