CodeOwnerSubmitRule: Fix exception logging
Rule errors are only returned in case an exception is caused by a user
misconfiguration (e.g. an invalid OWNERS file). In this case it is
sufficient to log a warning (without stacktrace).

For other exceptions CodeOwnerSubmitRule throws a
CodeOwnersInternalServerErrorException that will be logged on a higher
level. Hence CodeOwnerSubmitRule doesn't need to write an error log in
this case.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I2000893ce343d56073c2e10096eb6be8dfc9cd61