Add metric to record code owner config validations
Android would like to know how many pushes would be rejected if they
would turn on the code owner config validation. Collect this information
as a metric. Since the metric also covers the dry run mode, Android can
collect the stats with the dry run mode before turning on the validation
for real.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I257a9d2c3a13a603026f4da31120962700188c6e