CodeOwnerResolver: Use ImmutableList.Builder<String> to collect messages
CodeOwnerResolver consists out of several steps to resolve code owner
emails to code owner accounts. For each step we collect debug messages
so that issues with resolving code owners can easily be investigated.

So far the internal methods returned OptionalResultWithMessages which is
the result of the method plus collected debug messages. Going forward we
want to refactor CodeOwnerResolver to process all emails at once rather
than each email one by one. For this most of the steps will be converted
into operations on a stream, e.g. the methods will filter the stream or
map entries in the stream. When we do this the methods can no longer
return messages as part of the result.

To enable the refactoring, this change changes the internal methods to
no longer return OptionalResultWithMessages, but instead return the
plain result and get an ImmutableList.Builder<String> passed in that is
used to collect messages. This also simplifies some of the code.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Idf625ecf1784553dd6daa24a7d4862532e2524ac