CodeOwnerResolver: Split methods to lookup and check accounts
CodeOwnerResolver consists out of several steps to resolve code owner
emails to code owner accounts. At the moment we iterate over the code
owner emails and then execute these steps for each of the emails. Going
forward we want to refactor this so that all emails are processed at
once (so that each of the steps is executed only once, but with all
emails). For this most of the steps will be converted into operations on
a stream, e.g. the methods will filter the stream or map entries in the
stream.

To make this refactoring easier, this change splits the existing methods
that lookup accounts and check their visibility into smaller steps that
will become stream operations later.

The methods are sorted in the order in which the steps are invoked.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: If8e6686ebe4bd144c66d471739a9a39a72115a94