Fix parsing of globs that contain commas
per-file rules in OWNERS files allow to specify multiple globs as a
comma-separated list. The find-owners parser was naively splitting this
string by comma, which resulted in broken globs if a glob contained a
comma as part of a choice (e.g. "{foo,bar}) or as part of a character
class (e.g. "[,]").

Fix this by splitting the glob string only at commas that do not appear
withing curly or square brackets.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ia330dbbd82b42a416a7a6d1c4d24a2c5a81aae43