Make documentation of exemptedUser config more clear
If a user is exempted from requiring code owner approvals, it means that
patch sets that are uploaded by this user do not require code owner
approvals. If this user uploads a change, but another user uploaded the
last patch set for that change, the change is not exempted from
requiring code owner approvals.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ibeb695100c21b11480ba2e423d1a87c15680f599