Test submitting with implicit approval if REBASE_ALWAYS is used
When submitting a change it is required that the patch set that is
current at the moment when the Submit button is clicked has the
necessary code owner approvals. If a new patch set is created as part of
the submission (e.g. because REBASE_ALWAYS is used as a submit strategy)
that newly created patch set doesn't need to pass the code owner submit
check. Hence it is ok if this patch set doesn't have an implicit
approval. Verify from a test that this use case works.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I54ed7c8ee2b2c728096bdb73ece2c2ddbb43ca1d