CodeOwnerApprovalCheck: Remove wrong shortcut for fallback code owners
getFileStatusesForAccount(...) returned all files as approved if project
owners were configured as fallback code owners and the given user was a
project owner. Since fallback code owners are only applied for files
for which no code owners have been defined, doing this was incorrect if
a file had explcit code owners defined, but the given user was not
amongst them.

Since this was only a shortcut, we can simply drop it and the full
computation will compute the correct status.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I686bf7340e59c6ff852459113d8a412903e82910