CodeOwnerApprovalCheck: Do not check twice if a global owner is reviewer
getPathCodeOwnerStatus checks early whether any of the reviewers is a
global code owner and if yes sets the status to PENDING. In case
fallback code owners are evaluated we overrode that status which meant
that the fallback code owner logic had to check again whether any of the
reviewers was a global code owner. Refactor this so that the check is
done only once. This also improves the readability since the fallback
code owner logic now only needs to check about fallback code owners, and
no longer care about global code owners too.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I9bc8a28d6d7b6950e135abbd979e16641c9c829c