Parse annotations on code owners from OWNERS files
Extend the FindOwnersCodeOwnerConfigParser to be able to parse and
format annotations on email lines.

Annotations may appear at the end of an email line and have the
format: #{<annotation>}, e.g.:

  foo.bar@example.com #{NEVER_SUGGEST}

Email lines may have multiple annotations, e.g.:

  foo.bar@example.com #{FOO} #{BAR}

Other inline comments are still possible, e.g.:

  foo.bar@example.com #{FOO} #{BAR} # normal comment

The parsed annotations are stored as mutlimap of CodeOwnerReference to
CodeOwnerAnnotation in CodeOwnerSet. Annotations are represented as a
separate class, rather than a String, so that we can easily support
annotations values later, e.g. #{FOO=bar}.

With this change annotations are only support on email lines that
represent folder code owners, not on per-file lines that represent
per-file code owners.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I32249d77f2b4f8b7304c17dc786e402a98434d36