Make annotations on code owners available through CodeOwnerResolverResult
Annotations in OWNERS files are set on emails aka CodeOwnerReferences.
When resolving CodeOwnerReferences to CodeOwners, we must convert the
the multimap that stores the annotations per CodeOwnerReference into a
multimap that stores the annotations per CodeOwner (since we cannot
easily convert a CodeOwner back to a CodeOwnerReference in order to
lookup the annotations, e.g. we do not know which of their emails was
used as a reference).

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I06e902a50df4ba8b1909d67b2960992bf7ae1428